refactor(content-scripts): 重构内容脚本模块结构
将内容脚本相关功能迁移至 src/content-scripts 目录下，统一管理文本处理、
单词匹配、DOM 操作等核心逻辑。同时优化模块导出结构，提升代码可维护性。
```
