+++ src/components/transline/TransLine.tsx
@@ -1,4 +1,3 @@
-import type React from 'react';
 import T2 from './T2';
 
 /**
