+++ src/content-scripts/core.ts
