+++ src/content-scripts/textProcessor.ts
