+++ src/content-scripts/themeDetector.ts
